+++ changelog.md
@@ -1,3 +1,8 @@
+#dev-master
+
+ - split the scraping into an own repository: [paslandau\query-scraper](https://github.com/paslandau/query-scraper)
+ - caution: breaking changes to the API!
+
 ## 0.1.1
 
  - updated repositories to local satis installation